@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.api.objects.Update;
 import ru.ikusov.training.telegrambot.botservices.BotUpdateHandler;
-import ru.ikusov.training.telegrambot.repository.DatabaseConnector;
+import ru.ikusov.training.telegrambot.dao.DatabaseConnector;
 
 import java.util.List;
 
