@@ -5,7 +5,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
-import ru.ikusov.training.telegrambot.rabbit.RabbitPublisher;
 
 import java.util.Locale;
 import java.util.Optional;
@@ -17,26 +16,22 @@ public class MainClass {
     public static final String GEOCODE_API_KEY = Optional.ofNullable((System.getenv("geocode_api_key"))).orElse("");
     public static final Locale RUS_LOCALE = new Locale("ru", "RU");
 
-    public static void main(String... nomatterhowtocallthisargument) {
+    public static final boolean IS_TEST_MODE = false;
+//    public static final Long TEST_CHAT_ID = -467690956L;
+    public static final Long TEST_CHAT_ID = 349513007L;
 
-        //        //for proxy if needed (no)
-//        System.getProperties().put( "proxySet", "true" );
-//        System.getProperties().put( "socksProxyHost", "127.0.0.1" );
-//        System.getProperties().put( "socksProxyPort", "9150" );
+    public static void main(String... args) {
 
-        try (AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(MyConfig.class)) {
-            TelegramBotsApi botsApi;
-            Bot myBot;
-            myBot = context.getBean("bot", Bot.class);
-
-            botsApi = new TelegramBotsApi(DefaultBotSession.class);
+        try (var context = new AnnotationConfigApplicationContext(MyConfig.class)) {
+            log.info("Trying to get bot bean from {}...", context.getClass().getSimpleName());
+            var myBot = context.getBean("bot", Bot.class);
+            log.info("Bot bean successfully getted! Trying to create new TelegramBotsApi with DefaultBotSession...");
+            var botsApi = new TelegramBotsApi(DefaultBotSession.class);
+            log.info("TelegramBotsApi successfully created! Trying to register the bot in the api...");
             botsApi.registerBot(myBot);
-
-            log.info("Bot successfully registered.");
-//            new RabbitPublisher().start();
+            log.info("The Bot has been successfully registered!");
         } catch (Exception e) {
-            log.error("General Error: {}! Message was: {}", e.getClass(), e.getMessage());
-            e.printStackTrace();
+            log.error("General Error!", e);
         }
     }
 
