@@ -1,11 +1,15 @@
 package ru.ikusov.training.telegrambot.botservices;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import ru.ikusov.training.telegrambot.Bot;
+import ru.ikusov.training.telegrambot.services.UserNameGetter;
 
 public abstract class NonCommandMessageHandler implements MessageHandler {
+    protected final Logger log = LoggerFactory.getLogger(this.getClass());
+
     protected Bot bot;
 
     @Autowired
@@ -17,7 +21,19 @@ public abstract class NonCommandMessageHandler implements MessageHandler {
     public void handleMessage(Message message) {
         BotReaction botReaction = handleNonCommand(message);
         botReaction.react(bot);
+        botReaction.log();
     }
 
     public abstract BotReaction handleNonCommand(Message message);
+
+    protected void log(Message message) {
+        log.info(
+                "NON COMMAND ChatId: '{}' ChatName: '{}' UserId: '{}' UserName:'{}' Text: {}",
+                message.getChatId(),
+                message.getChat().getTitle(),
+                message.getFrom().getId(),
+                UserNameGetter.getUserName(message.getFrom()),
+                message.getText()
+        );
+    }
 }
