@@ -4,14 +4,25 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 import ru.ikusov.training.telegrambot.model.MyBotCommand;
 import ru.ikusov.training.telegrambot.utils.MessageType;
-import ru.ikusov.training.telegrambot.utils.RandomMessageGenerator;
 
 import java.util.Set;
 
 @Component
 @Order
-public class UnknownCommandMessageHandler extends CommandMessageHandler{
+public class UnknownCommandMessageHandler extends CommandMessageHandler {
     private final Set<String> commandVariants = Set.of();
+
+    @Override
+    public BotReaction handleCommand(MyBotCommand command) {
+        if (registeredCommands.contains(command.getCommand())) return new BotEmptyReaction();
+
+        super.log();
+        String textAnswer =
+                String.format(
+                        MessageType.UNKNOWN_COMMAND_MESSAGE.getRandomMessage(),
+                        command.getCommand());
+        return new BotMessageSender(command.getChatId(), textAnswer);
+    }
 
     @Override
     protected Set<String> getCommandVariants() {
@@ -24,16 +35,5 @@ public class UnknownCommandMessageHandler extends CommandMessageHandler{
     }
 
     @Override
-    public BotReaction handleCommand(MyBotCommand command) {
-        if(registeredCommands.contains(command.getCommand())) return new BotEmptyReaction();
-
-        String textAnswer =
-//                RandomMessageGenerator
-//                .generate(RandomMessageGenerator.MessageType.UNKNOWN_COMMAND_MESSAGE,
-//                        command.getCommand());
-                String.format(
-                        MessageType.UNKNOWN_COMMAND_MESSAGE.getRandomMessage(),
-                        command.getCommand());
-        return new BotMessageSender(command.getChatId(), textAnswer);
-    }
+    protected void log() {}
 }
