@@ -1,11 +1,11 @@
 package ru.ikusov.training.telegrambot.model;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import ru.ikusov.training.telegrambot.utils.MyString;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 
